# doc/EZshopBackEndApplication_swagger.yaml
openapi: 3.0.3
info:
  title: EZShop Back End Application API
  version: 1.0.0
  description: |
    OpenAPI definition for the EZShop backend server application.

    **EZShop Description**

    EZShop is a retail management system designed to handle the end-to-end operations of a small to medium-sized shop, integrating user management, inventory control, sales, orders, and accounting functions into a single platform.
    EZShop defines a set of business operations that orchestrates user interactions, product lifecycle, and financial transactions.

    In a Python-based reimplementation using FastAPI, EZShop would act as a RESTful backend service, exposing endpoints to manage authentication, product catalog, sales transactions, customer information, and payments.
    Authentication would rely on JWT-based bearer tokens, with access permissions differentiated by user roles (Administrator, ShopManager, Cashier).
    
    **Hierarchical Structure and Components**

    The system architecture is logically divided into four layers:

    1. Users and Authentication Layer
       - Manages user accounts and roles (Administrator, Cashier, ShopManager).
       - Provides endpoints for login, logout, and user CRUD operations.
       - Handles session security and authorization for all operations.

    2. Inventory and Product Management Layer
       - Handles product types, quantities, and barcodes (with optional RFID tracking).
...

...

...
      
...
    
...

...

...
servers:
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - name
      properties:
        code: 
          type: integer
          description: HTTP status code of the error
        message:
          type: string
          description: message of the error
        name:
          type: string
          description: Descriptive error message
      description: Standard structure for error responses
...
    ProductTypeUpdate:
      type: object
      properties:
        description:
          type: string
          minLength: 1
        barcode:
          type: string
          minLength: 1          
        price_per_unit:
          type: number
        note:
          type: string
          minLength: 1
        quantity:
          type: integer
        position:
          type: string
      required:
        - barcode
    ProductType:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
          minLength: 1
        barcode:
          type: string
          minLength: 1          
        price_per_unit:
          type: number
        note:
          type: string
          minLength: 1
        quantity:
          type: integer
        position:
          type: string
      required:
        - description
        - barcode
        - price_per_unit
    BooleanResponse:
      type: object
      properties:
        success: 
          type: boolean
      required: [success]
...
paths:
...
  /products:
    post:
      tags:
      - products
      summary: Create a new product type
      operationId: createProductType
      security:
        - bearerAuth: []
      description: |
        Create a new Product Type. Requires **Administrator** or **ShopManager** role. The success response is the product with its assigned ID.
        The barcode must be a string of 12-14 digits. Description and price per unit must be present. 
        Price must be > 0. If an id is provided in the request, it will be ignored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductType'
            
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductType'
              example:
                id: 1
                description: 'Chocolate Bar'
                barcode: '0123456780'
                price_per_unit: 2.99
                note: 'Imported from Belgium'
                quantity: 0
                position: '1-A-3'
        '400':
          description: Invalid input data (barcode must be a string of 12-14 digits). Description and price per unit must be present
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Bad request"
                name: "BadRequestError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 403
                message: "Insufficient rights"
                name: "ForbiddenError"
        '409':
          description: Product with same barcode already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 409
                message: "Insufficient rights"
                name: "ConflictError"
                
    get:
      tags: 
      - products
      summary: List all product types
      security:
      - bearerAuth: []
      description: Returns all the product type present in the shop. Requires **Administrator**, **ShopManager**, or **Cashier** role.
      responses:
        '200':
          description: List of product types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductType'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
  /products/{id}:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
    get:
      tags:
      - products
      summary: Get product type by ID
      security:
        - bearerAuth: []
      description: Return a product type by its ID. Requires **Administrator**, **ShopManager**, or **Cashier** role. 
      operationId: getProductType
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductType'
        '400':
          description: |
            Missing id or invalid id. ID must be a positive integer.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Bad request"
                name: "BadRequestError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 403
                message: "Insufficient rights"
                name: "ForbiddenError"
        '404':
          description: Product Type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Product not found"
                name: "NotFoundError"
    put:
      tags:
      - products
      summary: Update product type by ID
      operationId: updateProduct
      security:
        - bearerAuth: []
      description: |
        Update product type by ID. Requires **Administrator**, **ShopManager** role. Price and quantity, if present, must be > 0.
        A product type barcode cannot be updated if the new value is already assigned to another product type.
        Barcode can be updated only if: 
          - The product has NO sales associated with it. 
          - The product has NO orders associated with it.
          - The product has NO returns associated with it.

        In case the price is updated in product type, the eventual open sale will not be affected, as it will count the "old" price. 
        Return amount will refer to the paid amount (so the one in the sale), not the current price of product type.
      requestBody:
        requi1red: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductTypeUpdate'
      responses:
        '201':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponse'
        '400':
          description: |
            Invalid input data (barcode must be a string of 12-14 digits). 
            Quantity and price per unit, if present, must be positive number.
            Product Id must be a valid product id ( > 0 ). 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Bad request"
                name: "BadRequestError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 403
                message: "Insufficient rights"
                name: "ForbiddenError"
        '404':
          description: Product Type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Product not found"
                name: "NotFoundError"
        '409':
          description: Barcode/position already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 409
                message: "Barcode already in use"
                name: "ConflictError"
        '420':
          description: Invalid order/sale/return state prevents barcode update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 420
                message: "Invalid sale state"
                name: "InvalidStateError"
    delete:
      tags:
      - products
      summary: Delete product type by ID
      operationId: deleteProductType
      security:
        - bearerAuth: []
      description: |
        Delete product type by ID. Requires **Administrator**, **ShopManager** role.
        A product in a sale (OPEN or PENDING or PAID) cannot be deleted.
        A product in an order (ISSUED or PAID or COMPLETED) cannot be deleted.
        A product in a return (OPEN or CLOSED or REIMBURSED) cannot be deleted.
      responses:
        '204':
          description: Deleted
        '400':
          description: Invalid product id ( <= 0 )
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Invalid product id"
                name: "BadRequestError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 403
                message: "Insufficient rights"
                name: "ForbiddenError"
        '404':
          description: Product Type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Product not found"
                name: "NotFoundError"
        '420':
          description: Invalid order/sale/return state prevents deletion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 420
                message: "Invalid sale state"
                name: "InvalidStateError"
  /products/barcode/{barcode}:
    parameters:
    - name: barcode
      in: path
      required: true
      schema:
        type: string
    get:
      tags:
      - products
      summary: Get product by barcode
      security:
        - bearerAuth: []
      description: Return product by barcode. Requires **Administrator** or **ShopManager** role.
      operationId: getProductTypeByBarCode
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductType'
        '400':
          description: Invalid barcode or barcode format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Invalid barccode"
                name: "BadRequestError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 403
                message: "Insufficient rights"
                name: "ForbiddenError"
        '404':
          description: Product Type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Product not found"
                name: "NotFoundError"
  /products/search:
    get:
      tags:
      - products
      summary: Search product types by description
      operationId: getProductTypesByDescription
      security:
        - bearerAuth: []
      description: Search product types by description. Requires **Administrator** or **ShopManager** role.
      parameters:
      - name: query
        in: query
        required: false
        schema:
          type: string
        description: Partial description to filter products
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductType'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 403
                message: "Insufficient rights"
                name: "ForbiddenError"
  /products/{id}/position:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
    patch:
      tags:
      - products
      summary: Assign or clear shelf position (aisle-shelf-level).
      operationId: updatePosition
      security:
        - bearerAuth: []
      description: Requires **Administrator** or **ShopManager** role.  
        The position must match pattern **<digits>-<letters>-<digits>**. To clear the position, send an empty string.
      parameters:
        - name: position
          in: query
          required: true
          schema:
            type: string
            example: "2-B-1"
      responses:
        '201':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponse'
        '400':
          description: Invalid product id or invalid position string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Invalid product id"
                name: "BadRequestError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 403
                message: "Insufficient rights"
                name: "ForbiddenError"
        '404':
          description: Product Type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Product not found"
                name: "NotFoundError"
        '409':
          description: Position conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 409
                message: "Position conflict"
                name: "ConflictError"
  /products/{id}/quantity:
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: integer
    patch:
      tags:
      - products
      summary: Increase/decrease available quantity. 
      description: Modify the available quantity. A negative number decrease the available quantity. Requires **Administrator** or **ShopManager** role.
      operationId: updateQuantity
      parameters:
        - name: quantity
          in: query
          required: true
          schema:
            type: integer
            example: 5
      responses:
        '201':
          description: Updated quantity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponse'
        '400':
          description: Invalid product id, invalid or insufficient quantity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: "Invalid product id"
                name: "BadRequestError"
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 401
                message: "Authorization header missing"
                name: "UnauthorizedError"
        '403':
          description: Unauthorized (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 403
                message: "Insufficient rights"
                name: "ForbiddenError"
        '404':
          description: Product Type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Product not found"
                name: "NotFoundError"
...